Updates users-model and users-router
These files have been updated - i'm trying to add a getByUsername endpoint, for authentication purposes. Still trying to get things in place to get passport-local up and running.

In the past commit I implemented bcrypt and updated the register endpoint. Now when a user registers their passwords are hashed in the database. Tested endpoint in Insomnia. ✅✅✅✅