Adds register with passport
I'm stuck, but a little less stuck than I was this morning.

The passport-config seems to be doing *some* of the stuff I expect it to. A user can succesfully register themselves through the /register endpoint. A GET takes you to an ejs register form, fill out the form and hit submit, and boom - a new user with a hashed password is added to the database.

The problem I'm still having is logging in.....

If a user gives the right username and the password is the same (after being hashed again, which should happen through the passport-config, in the try block of the initialize function). Then a user should just get redirected to / - or index.ejs, which is nothing, just a Hi message.

The problem lies in the getUser and getUserById functions - I think... I need to approach these differently, but I don't know how.